Fix warnings found by compiling with MS VC++ 2019
This change resolves all found warnings including 'nodiscard'
warnings mentioned in the issue description.

Components: Vulkan, OpenGL, EGL

VK-GL-CTS public issue: 188

Affects:
*

+++ external/openglcts/modules/common/glcPackedPixelsTests.cpp
@@ -948,7 +948,7 @@ void RectangleTest::applyInitialStorageModes()
 
 void RectangleTest::swapBytes(int typeSize, std::vector<GLbyte>& dataBuffer)
 {
-	int bufferSize = dataBuffer.size();
+	int bufferSize = static_cast<int>(dataBuffer.size());
 	switch (typeSize)
 	{
 	case 1:
